# cargo-ship-backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/monitoring/data": {
      "post": {
        "description": "接收并存储单条设备监测数据",
        "operationId": "MonitoringController_receiveMonitoringData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimeSeriesDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "接收成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "dataId": {
                          "type": "number",
                          "example": 123456
                        },
                        "received": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "timestamp": {
                      "type": "number",
                      "example": 1700000000000
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "设备不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "接收单条监测数据",
        "tags": [
          "监测数据"
        ]
      },
      "get": {
        "description": "根据设备、指标类型、时间范围查询监测数据，支持分页",
        "operationId": "MonitoringController_queryMonitoringData",
        "parameters": [
          {
            "name": "equipmentId",
            "required": true,
            "in": "query",
            "description": "设备唯一标识（UUID格式）",
            "schema": {
              "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
              "type": "string"
            }
          },
          {
            "name": "metricType",
            "required": false,
            "in": "query",
            "description": "指标类型（可选）",
            "schema": {
              "example": "vibration",
              "type": "string",
              "enum": [
                "vibration",
                "temperature",
                "pressure",
                "humidity",
                "speed",
                "current",
                "voltage",
                "power"
              ]
            }
          },
          {
            "name": "startTime",
            "required": true,
            "in": "query",
            "description": "开始时间戳（毫秒）",
            "schema": {
              "example": 1700000000000,
              "type": "number"
            }
          },
          {
            "name": "endTime",
            "required": true,
            "in": "query",
            "description": "结束时间戳（毫秒）",
            "schema": {
              "example": 1700086400000,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码（从1开始）",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 100,
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "total": {
                          "type": "number",
                          "example": 5000
                        },
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "pageSize": {
                          "type": "number",
                          "example": 100
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 50
                        }
                      }
                    },
                    "timestamp": {
                      "type": "number",
                      "example": 1700000000000
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "设备不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询监测数据",
        "tags": [
          "监测数据"
        ]
      }
    },
    "/api/monitoring/data/batch": {
      "post": {
        "description": "批量接收并存储设备监测数据，最多1000条/次",
        "operationId": "MonitoringController_receiveBatchMonitoringData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchTimeSeriesDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "接收成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalCount": {
                          "type": "number",
                          "example": 100
                        },
                        "successCount": {
                          "type": "number",
                          "example": 98
                        },
                        "failedCount": {
                          "type": "number",
                          "example": 2
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "index": {
                                "type": "number",
                                "example": 5
                              },
                              "reason": {
                                "type": "string",
                                "example": "数据格式错误"
                              }
                            }
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "number",
                      "example": 1700000000000
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "设备不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量接收监测数据",
        "tags": [
          "监测数据"
        ]
      }
    },
    "/api/monitoring/data/statistics": {
      "get": {
        "description": "获取指定设备、指标类型、时间范围的统计数据（最大值、最小值、平均值等）",
        "operationId": "MonitoringController_getDataStatistics",
        "parameters": [
          {
            "name": "equipmentId",
            "required": true,
            "in": "query",
            "description": "设备ID（UUID格式）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metricType",
            "required": true,
            "in": "query",
            "description": "指标类型",
            "schema": {
              "enum": [
                "vibration",
                "temperature",
                "pressure",
                "humidity",
                "speed",
                "current",
                "voltage",
                "power"
              ],
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "required": true,
            "in": "query",
            "description": "开始时间戳（毫秒）",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "endTime",
            "required": true,
            "in": "query",
            "description": "结束时间戳（毫秒）",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "metricType": {
                          "type": "string",
                          "example": "vibration"
                        },
                        "count": {
                          "type": "number",
                          "example": 1000
                        },
                        "maxValue": {
                          "type": "number",
                          "example": 25.5
                        },
                        "minValue": {
                          "type": "number",
                          "example": 8.2
                        },
                        "avgValue": {
                          "type": "number",
                          "example": 15.3
                        },
                        "unit": {
                          "type": "string",
                          "example": "mm/s"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "number",
                      "example": 1700000000000
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "设备不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取数据统计信息",
        "tags": [
          "监测数据"
        ]
      }
    },
...
  },
  "info": {
    "title": "TDuCargoShipsManagement API",
    "description": "船舶货物管理系统 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTimeSeriesDataDto": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "string",
            "description": "设备唯一标识（UUID格式）",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "description": "数据时间戳",
            "example": "2025-11-19T10:30:00.000Z"
          },
          "metricType": {
            "type": "string",
            "description": "指标类型",
            "enum": [
              "vibration",
              "temperature",
              "pressure",
              "humidity",
              "speed",
              "current",
              "voltage",
              "power"
            ],
            "example": "vibration"
          },
          "value": {
            "type": "number",
            "description": "指标数值",
            "example": 12.5,
            "minimum": -999999.99,
            "maximum": 999999.99
          },
          "unit": {
            "type": "string",
            "description": "数据单位（可选，未提供时使用默认值）",
            "example": "mm/s",
            "maxLength": 20
          },
          "quality": {
            "type": "string",
            "description": "数据质量标记（可选，默认为normal）",
            "enum": [
              "normal",
              "abnormal",
              "suspicious"
            ],
            "default": "normal"
          },
          "source": {
            "type": "string",
            "description": "数据来源（可选，默认为sensor-upload）",
            "enum": [
              "sensor-upload",
              "file-import",
              "manual-entry"
            ],
            "default": "sensor-upload"
          }
        },
        "required": [
          "equipmentId",
          "timestamp",
          "metricType",
          "value"
        ]
      },
      "BatchDataItemDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "description": "数据时间戳",
            "example": "2025-11-19T10:30:00.000Z"
          },
          "metricType": {
            "type": "string",
            "description": "指标类型",
            "enum": [
              "vibration",
              "temperature",
              "pressure",
              "humidity",
              "speed",
              "current",
              "voltage",
              "power"
            ],
            "example": "vibration"
          },
          "value": {
            "type": "number",
            "description": "指标数值",
            "example": 12.5,
            "minimum": -999999.99,
            "maximum": 999999.99
          },
          "unit": {
            "type": "string",
            "description": "数据单位（可选）",
            "example": "mm/s"
          },
          "quality": {
            "type": "string",
            "description": "数据质量标记（可选）",
            "enum": [
              "normal",
              "abnormal",
              "suspicious"
            ]
          }
        },
        "required": [
          "timestamp",
          "metricType",
          "value"
        ]
      },
      "CreateBatchTimeSeriesDataDto": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "string",
            "description": "设备唯一标识（UUID格式）",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "data": {
            "description": "监测数据数组（1-1000条）",
            "minItems": 1,
            "maxItems": 1000,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchDataItemDto"
            }
          }
        },
        "required": [
          "equipmentId",
          "data"
        ]
      },
...
    }
  }
}